requirements.txt
boto3~=1.7.0    
python-dateutil~=2.7.0     
